Preserve checkpoint dtype on load; add checkpoint diagnostic
load_cache_head built CacheHead in float32 and then called load_state_dict,
which copies into the existing parameters rather than adopting the stored
dtype. A bf16 checkpoint therefore came back as float32. The inference runner
compounded it by never casting the loaded head, unlike the no-checkpoint branch
right beside it which does .to(device=device, dtype=dtype).

The consequence is silent rather than loud: the float32 head emits float32
tokens, scheduler.step promotes the latents to float32 (both operands are
dimensioned, so bf16 does not win), and the next full step hands float32
activations to a bf16 Wan.

load_cache_head now takes a dtype, and run_pipeline passes the pipeline dtype.

Also adds diagnose_checkpoints.py, for the case where two checkpoints produce
byte-identical video. It reports whether out_proj is still zero-init (the head
never left carry_previous), whether the weights differ at all, and whether the
residual survives the bf16 add in head_step -- a residual below roughly 0.2% of
the token magnitude is largely rounded away, so a head can be training and
still produce identical output.

The first of those is the expected early state of --arm dmd --warmup-steps 0:
FakeScoreWan is zero-delta at init, so fake_x0 == teacher_x0 exactly, dmd_loss
returns 0, and the head receives no gradient at all until the fake-score LoRA
has learned to differ from the teacher.

+++ examples/wanvideo/model_inference/cache_head/cache_head_model_inference.py
@@ -196,7 +196,7 @@ def run_pipeline(args) -> None:
     # Head + config.
     config = None
     if args.checkpoint and Path(args.checkpoint).is_file():
-        head, config = load_cache_head(args.checkpoint, device=device)
+        head, config = load_cache_head(args.checkpoint, device=device, dtype=dtype)
         print(f"Loaded CacheHead from {args.checkpoint} (cfg={config.cfg_scale})")
     else:
         config = CacheHeadConfig(model_id=args.model_id, cfg_scale=args.cfg)
